refactor(moyo): gather layer identify modules into identify/layer
Move `identify/layer_group.rs` and `identify/layer_point_group.rs`
into a new `identify/layer/` submodule, alongside a new
`identify/layer/normalizer.rs` that holds the layer-specific
`integral_normalizer_2_1` (with its `layer_exact_s_z` helper) split
out of the bulk `identify/normalizer.rs`. The bulk file now contains
only `integral_normalizer`.

The new `moyo/src/identify/layer.rs` declares the three submodules as
`pub(super) mod` and re-exports `LayerGroup`, `LayerPointGroup`, and
`integral_normalizer_2_1`. `identify.rs` switches the corresponding
re-exports to go through `layer::*`. The public API surface
(`crate::identify::LayerGroup`, `crate::identify::LayerPointGroup`,
`crate::identify::integral_normalizer_2_1`,
`crate::identify::integral_normalizer`) is unchanged.

Internal `super::*` paths inside `layer/layer_point_group.rs` and
the new `layer/normalizer.rs` become `super::super::*` for top-level
identify siblings (`point_group`, `rotation_type`, `space_group`).
Intra-`layer/` siblings (`layer_group` -> `layer_point_group`,
`layer_group` -> `normalizer`) keep the simple `super::*` form.

No behavior change; all 178 cargo tests pass.
