+++ moyo/src/identify.rs
@@ -1,15 +1,13 @@
-mod layer_group;
-mod layer_point_group;
+mod layer;
 mod magnetic_space_group;
 mod normalizer;
 mod point_group;
 mod rotation_type;
 mod space_group;
 
-pub use layer_group::LayerGroup;
-pub use layer_point_group::LayerPointGroup;
+pub use layer::{LayerGroup, LayerPointGroup, integral_normalizer_2_1};
 pub use magnetic_space_group::MagneticSpaceGroup;
-pub use normalizer::{integral_normalizer, integral_normalizer_2_1};
+pub use normalizer::integral_normalizer;
 pub use point_group::PointGroup;
 pub use space_group::SpaceGroup;
